@@ -22,34 +22,20 @@ import android.app.Activity;
 import android.app.AlertDialog;
 import android.app.Dialog;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.content.res.Configuration;
 import android.graphics.*;
-import android.hardware.camera2.CameraAccessException;
-import android.hardware.camera2.CameraCaptureSession;
-import android.hardware.camera2.CameraCharacteristics;
-import android.hardware.camera2.CameraDevice;
-import android.hardware.camera2.CameraManager;
-import android.hardware.camera2.CameraMetadata;
-import android.hardware.camera2.CaptureRequest;
-import android.hardware.camera2.CaptureResult;
-import android.hardware.camera2.TotalCaptureResult;
+import android.hardware.camera2.*;
+import android.hardware.camera2.params.ColorSpaceTransform;
+import android.hardware.camera2.params.MeteringRectangle;
 import android.hardware.camera2.params.StreamConfigurationMap;
-import android.media.Image;
 import android.media.ImageReader;
+import android.media.MediaPlayer;
+import android.net.Uri;
 import android.os.*;
 import android.util.*;
-import android.view.LayoutInflater;
-import android.view.Surface;
-import android.view.TextureView;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.ImageButton;
-import android.widget.ImageView;
-import android.widget.ProgressBar;
+import android.view.*;
 import android.widget.Toast;
 import android.widget.ToggleButton;
 import androidx.annotation.NonNull;
@@ -60,30 +46,28 @@ import androidx.fragment.app.DialogFragment;
 import androidx.fragment.app.Fragment;
 import com.eszdman.photoncamera.AutoFitTextureView;
 import com.eszdman.photoncamera.Parameters.ExposureIndex;
-import com.eszdman.photoncamera.R;
-import com.eszdman.photoncamera.api.CameraManager2;
-import com.eszdman.photoncamera.api.CameraReflectionApi;
 import com.eszdman.photoncamera.Parameters.FrameNumberSelector;
 import com.eszdman.photoncamera.Parameters.IsoExpoSelector;
+import com.eszdman.photoncamera.R;
+import com.eszdman.photoncamera.SurfaceViewOverViewfinder;
+import com.eszdman.photoncamera.api.CameraManager2;
+import com.eszdman.photoncamera.api.CameraReflectionApi;
 import com.eszdman.photoncamera.api.ImageSaver;
 import com.eszdman.photoncamera.api.Interface;
 import com.eszdman.photoncamera.gallery.GalleryActivity;
+import com.eszdman.photoncamera.util.CustomLogger;
+import com.eszdman.photoncamera.util.FileManager;
+import rapid.decoder.BitmapDecoder;
 
 import java.io.File;
-import java.io.FilenameFilter;
 import java.lang.reflect.Field;
 import java.util.*;
 import java.util.concurrent.Semaphore;
 import java.util.concurrent.TimeUnit;
 
-import com.eszdman.photoncamera.util.CustomLogger;
-import com.eszdman.photoncamera.util.FileManager;
-import com.eszdman.photoncamera.util.Utilities;
-import de.hdodenhof.circleimageview.CircleImageView;
-
 import static androidx.constraintlayout.widget.ConstraintSet.WRAP_CONTENT;
-import static com.eszdman.photoncamera.ui.MainActivity.onCameraViewCreated;
 
+@SuppressWarnings("rawtypes")
 public class CameraFragment extends Fragment
         implements View.OnClickListener, ActivityCompat.OnRequestPermissionsResultCallback {
 
@@ -157,6 +141,7 @@ public class CameraFragment extends Fragment
     public long mPreviewExposuretime;
     public int mPreviewIso;
     public Rational[] mPreviewTemp;
+    public ColorSpaceTransform mColorSpaceTransform;
     Range FpsRangeDef;
     Range FpsRangeHigh;
     private float mFocus;
@@ -168,22 +153,23 @@ public class CameraFragment extends Fragment
             = new TextureView.SurfaceTextureListener() {
 
         @Override
-        public void onSurfaceTextureAvailable(SurfaceTexture texture, int width, int height) {
+        public void onSurfaceTextureAvailable(@NonNull SurfaceTexture texture, int width, int height) {
             openCamera(width, height);
         }
 
         @Override
-        public void onSurfaceTextureSizeChanged(SurfaceTexture texture, int width, int height) {
+        public void onSurfaceTextureSizeChanged(@NonNull SurfaceTexture texture, int width, int height) {
             configureTransform(width, height);
         }
 
         @Override
-        public boolean onSurfaceTextureDestroyed(SurfaceTexture texture) {
+
+        public boolean onSurfaceTextureDestroyed(@NonNull SurfaceTexture texture) {
             return true;
         }
 
         @Override
-        public void onSurfaceTextureUpdated(SurfaceTexture texture) {
+        public void onSurfaceTextureUpdated(@NonNull SurfaceTexture texture) {
         }
 
     };
@@ -198,6 +184,7 @@ public class CameraFragment extends Fragment
      * An {@link AutoFitTextureView} for camera preview.
      */
     public AutoFitTextureView mTextureView;
+    public SurfaceViewOverViewfinder surfaceView;
 
     /**
      * A {@link CameraCaptureSession } for camera preview.
@@ -416,8 +403,9 @@ public class CameraFragment extends Fragment
             Rational[] mtemp = result.get(CaptureResult.SENSOR_NEUTRAL_COLOR_POINT);
             if(exposure != null) mPreviewExposuretime = (long)exposure;
             if(iso != null) mPreviewIso = (int)iso;
-            if(focus != null) mFocus = (float)focus; else focus = 0.f;
+            if(focus != null) mFocus = (float)focus;
             mPreviewTemp = mtemp;
+            mColorSpaceTransform = result.get(CaptureResult.COLOR_CORRECTION_TRANSFORM);
             process(result);
             updateScreenLog(result);
         }
@@ -436,7 +424,7 @@ public class CameraFragment extends Fragment
                     }
                 }
                 if(ExposureIndex.index()+0.9 < 8.0) {
-                    if(is30Fps && Interface.i.settings.fpsPreview && mCameraDevice.getId() != "1")
+                    if(is30Fps && Interface.i.settings.fpsPreview && !mCameraDevice.getId().equals("1"))
                     {
                         Log.d(TAG,"Changed preview target 60fps");
                         mPreviewRequestBuilder.set(CaptureRequest.CONTROL_AE_TARGET_FPS_RANGE,FpsRangeHigh);
@@ -452,20 +440,43 @@ public class CameraFragment extends Fragment
 
     void updateScreenLog(CaptureResult result) {
         CustomLogger cl = new CustomLogger(getActivity(), R.id.screen_log_focus);
-        if (true) { //Preference to be added here
+        if (Interface.i.settings.afdata) {
+            IsoExpoSelector.ExpoPair expoPair = IsoExpoSelector.GenerateExpoPair(0);
             LinkedHashMap<String, String> dataset = new LinkedHashMap<>();
             dataset.put("AF_MODE", getResultFieldName("CONTROL_AF_MODE_", result.get(CaptureResult.CONTROL_AF_MODE)));
             dataset.put("AF_TRIGGER", getResultFieldName("CONTROL_AF_TRIGGER_", result.get(CaptureResult.CONTROL_AF_TRIGGER)));
             dataset.put("AF_STATE", getResultFieldName("CONTROL_AF_STATE_", result.get(CaptureResult.CONTROL_AF_STATE)));
             dataset.put("FOCUS_DISTANCE", String.valueOf(result.get(CaptureResult.LENS_FOCUS_DISTANCE)));
+            dataset.put("EXPOSURE_TIME", expoPair.ExposureString() + "s");
+            dataset.put("ISO", String.valueOf(expoPair.iso));
+            dataset.put("Shakeness", String.valueOf(Interface.i.sensors.getShakeness()));
             dataset.put("FOCUS_RECT", Arrays.deepToString(result.get(CaptureResult.CONTROL_AF_REGIONS)));
-            dataset.put("EXPOSURE_TIME", Utilities.formatExposureTime(Objects.requireNonNull(result.get(CaptureResult.SENSOR_EXPOSURE_TIME)).doubleValue() / 1E9) + "s");
-            dataset.put("ISO", String.valueOf(result.get(CaptureResult.SENSOR_SENSITIVITY)));
+            MeteringRectangle[] rectobj = result.get(CaptureResult.CONTROL_AF_REGIONS);
+            if(rectobj != null && rectobj.length > 0) {
+                RectF rect = getScreenRectFromMeteringRect(rectobj[0]);
+                dataset.put("F_RECT(px)", rect.toString());
+                surfaceView.update(rect);
+            }
             cl.setVisibility(View.VISIBLE);
             cl.updateText(cl.createTextFrom(dataset));
         } else {
+            surfaceView.update(null);
             cl.setVisibility(View.GONE);
         }
+    }
+
+    private RectF getScreenRectFromMeteringRect(MeteringRectangle meteringRectangle) {
+        if(mImageReaderPreview == null) return new RectF();
+            float left = (((float) meteringRectangle.getY() / mImageReaderPreview.getHeight()) * (surfaceView.getWidth()));
+            float top = (((float) meteringRectangle.getX() / mImageReaderPreview.getWidth())* (surfaceView.getHeight()));
+            float width = (((float) meteringRectangle.getHeight() / mImageReaderPreview.getHeight()) * (surfaceView.getWidth()));
+            float height = (((float) meteringRectangle.getWidth() / mImageReaderPreview.getWidth()) * (surfaceView.getHeight()));
+            return new RectF(
+                          left        , //Left
+                          top         ,  //Top
+                    left + width,//Right
+                    top+height //Bottom
+            );
     }
 
     private String getResultFieldName(String prefix, Integer value) {
@@ -489,12 +500,7 @@ public class CameraFragment extends Fragment
     public void showToast(final String text) {
         final Activity activity = getActivity();
         if (activity != null) {
-            activity.runOnUiThread(new Runnable() {
-                @Override
-                public void run() {
-                    Toast.makeText(activity, text, Toast.LENGTH_SHORT).show();
-                }
-            });
+            activity.runOnUiThread(() -> Toast.makeText(activity, text, Toast.LENGTH_SHORT).show());
         }
     }
 
@@ -605,17 +611,12 @@ public class CameraFragment extends Fragment
         return getAdjustedLayout(aspectRatio, activity_main);
     }
 
-    public ImageButton shot;
-    public ProgressBar lightcycle;
-    public static ProgressBar loadingcycle;
-    public CircleImageView galleryImageButton;
-
     @Override
     public void onClick(View view) {
         switch (view.getId()) {
             case R.id.picture: {
-                shot.setActivated(false);
-                shot.setClickable(false);
+                Interface.i.cameraui.shot.setActivated(false);
+                Interface.i.cameraui.shot.setClickable(false);
                 takePicture();
                 break;
             }
@@ -631,12 +632,11 @@ public class CameraFragment extends Fragment
                 if (sw.isChecked()) {
                     mTargetFormat = rawFormat;
                     Interface.i.settings.hdrx = true;
-                    Interface.i.settings.save();
                 } else {
                     mTargetFormat = yuvFormat;
                     Interface.i.settings.hdrx = false;
-                    Interface.i.settings.save();
                 }
+                Interface.i.settings.save();
                 restartCamera();
                 break;
             }
@@ -649,81 +649,9 @@ public class CameraFragment extends Fragment
 
     @Override
     public void onViewCreated(final View view, Bundle savedInstanceState) {
-        onCameraViewCreated();
-        Interface.i.touchFocus.ReInit();
-        lightcycle = view.findViewById(R.id.lightCycle);
-        lightcycle.setAlpha(0);
-        lightcycle.setMax(Interface.i.settings.frameCount);
-        //lightcycle.setMin(0);
-        loadingcycle = view.findViewById(R.id.progressloading);
-        loadingcycle.setMax(Interface.i.settings.frameCount);
-        shot = view.findViewById(R.id.picture);
-        shot.setOnClickListener(this);
-        shot.setActivated(true);
-        ToggleButton fpsPreview = view.findViewById(R.id.fpsPreview);
-        fpsPreview.setChecked(Interface.i.settings.fpsPreview);
-        ToggleButton quadResolution = view.findViewById(R.id.quadRes);
-        quadResolution.setChecked(Interface.i.settings.QuadBayer);
-        ToggleButton eisPhoto = view.findViewById(R.id.eisPhoto);
-        eisPhoto.setChecked(Interface.i.settings.eisPhoto);
-        eisPhoto.setOnClickListener(v -> {
-            Interface.i.settings.eisPhoto = !Interface.i.settings.eisPhoto;
-            Interface.i.settings.save();
-        });
-        fpsPreview.setOnClickListener(v -> {
-            Interface.i.settings.fpsPreview = !Interface.i.settings.fpsPreview;
-            Interface.i.settings.save();
-        });
-        quadResolution.setOnClickListener(v -> {
-            Interface.i.settings.QuadBayer = !Interface.i.settings.QuadBayer;
-            Interface.i.settings.save();
-            restartCamera();
-        });
-        ImageButton flip = view.findViewById(R.id.flip_camera);
-        flip.setOnClickListener(v -> {
-                flip.animate().rotation(flip.getRotation() - 180).setDuration(450).start();
-                Interface.i.settings.mCameraID = cycler(Interface.i.settings.mCameraID, mCameraIds);
-                Interface.i.settings.save();
-                restartCamera();
-            });
-        Button settings = view.findViewById(R.id.settings);
-        settings.setOnClickListener(this);
-        ToggleButton hdrX = view.findViewById(R.id.stacking);
-        hdrX.setOnClickListener(this);
-        galleryImageButton = view.findViewById(R.id.ImageOut);
-        galleryImageButton.setOnClickListener(this);
-        galleryImageButton.setClickable(true);
-        loadGalleryButtonImage();
-        ToggleButton night = view.findViewById(R.id.nightMode);
-        ToggleButton video = view.findViewById(R.id.videoMode);
-        ToggleButton camera = view.findViewById(R.id.cameraMode);
-        camera.setChecked(true);
-        night.setChecked(Interface.i.settings.nightMode);
-        night.setOnCheckedChangeListener((buttonView, isChecked) -> {
-            if(night.isChecked()){
-                camera.setChecked(false);
-                video.setChecked(false);
-            }
-            Interface.i.settings.nightMode = !Interface.i.settings.nightMode;
-            Interface.i.settings.save();
-            restartCamera();
-        });
-
-        camera.setOnCheckedChangeListener((buttonView, isChecked) -> {
-            if(camera.isChecked()){
-                night.setChecked(false);
-                video.setChecked(false);
-            }
-        });
-
-        video.setOnCheckedChangeListener((buttonView, isChecked) -> {
-            if(video.isChecked()){
-                night.setChecked(false);
-                camera.setChecked(false);
-            }
-        });
-
         mTextureView = view.findViewById(R.id.texture);
+        surfaceView = view.findViewById(R.id.surfaceView);
+        Interface.i.cameraui.onCameraViewCreated();
     }
 
     @Override
@@ -732,13 +660,8 @@ public class CameraFragment extends Fragment
 
     }
 
-    private void loadGalleryButtonImage() {
-        File[] files = FileManager.DCIM_CAMERA.listFiles(new FilenameFilter() {
-            @Override
-            public boolean accept(File dir, String name) {
-                return name.toUpperCase().endsWith(".JPG");
-            }
-        } );
+    public void loadGalleryButtonImage() {
+        File[] files = FileManager.DCIM_CAMERA.listFiles((dir, name) -> name.toUpperCase().endsWith(".JPG"));
         if (files != null) {
             long lastModifiedTime = -1;
             File lastImage = null;
@@ -748,31 +671,20 @@ public class CameraFragment extends Fragment
                     lastModifiedTime = f.lastModified();
                 }
             }
-            //TODO replace this with Bitmap.createScaledBitmap later
-            Bitmap bitmap = BitmapFactory.decodeFile(lastImage.getAbsolutePath());
-            galleryImageButton.setImageBitmap(bitmap);
+            //Used fastest decoder on the wide west
+            if(lastImage != null) {
+                    Interface.i.cameraui.galleryImageButton.setImageBitmap(
+                            BitmapDecoder.from(Uri.fromFile(lastImage))
+                                    .scaleBy(0.1f)
+                                    .decode());
+            }
         }
     }
 
     @Override
     public void onResume() {
         super.onResume();
-
-        Log.d(TAG,"CameraResume");
-        MainActivity.act.onCameraResume();
-        Interface.i.touchFocus.ReInit();
-        ImageView grid_icon = MainActivity.act.findViewById(R.id.grid);
-        ImageView edges = MainActivity.act.findViewById(R.id.edges);
-        ToggleButton hdrX = MainActivity.act.findViewById(R.id.stacking);
-        Interface.i.gravity.run();
-        if (Interface.i.settings.grid) grid_icon.setVisibility(View.VISIBLE);
-        else grid_icon.setVisibility(View.GONE);
-        if (Interface.i.settings.roundedge) edges.setVisibility(View.VISIBLE);
-        else edges.setVisibility(View.GONE);
-        hdrX.setChecked(Interface.i.settings.hdrx);
-        ToggleButton night = Interface.i.mainActivity.findViewById(R.id.nightMode);
-        night.setChecked(Interface.i.settings.nightMode);
-        startBackgroundThread();
+        Interface.i.cameraui.onCameraResume();
         if (mTextureView == null) mTextureView = new AutoFitTextureView(MainActivity.act);
         if (mTextureView.isAvailable()) {
             openCamera(mTextureView.getWidth(), mTextureView.getHeight());
@@ -784,9 +696,8 @@ public class CameraFragment extends Fragment
 
     @Override
     public void onPause() {
-        Interface.i.gravity.stop();
+        Interface.i.cameraui.onCameraPause();
         closeCamera();
-        Interface.i.settings.saveID();
         stopBackgroundThread();
         super.onPause();
     }
@@ -836,7 +747,9 @@ public class CameraFragment extends Fragment
              target = sizes.get(s);
              if(Interface.i.settings.QuadBayer) {
                  Rect pre = mCameraCharacteristics.get(CameraCharacteristics.SENSOR_INFO_PRE_CORRECTION_ACTIVE_ARRAY_SIZE);
+                 if(pre == null) return target;
                  Rect act = mCameraCharacteristics.get(CameraCharacteristics.SENSOR_INFO_ACTIVE_ARRAY_SIZE);
+                 if(act == null) return target;
                  double k = (double) (target.getHeight()) / act.bottom;
                  mul(pre, k);
                  mul(act, k);
@@ -853,24 +766,21 @@ public class CameraFragment extends Fragment
          }
          return mPreviewSize;
      }
+
+    int mPreviewwidth;
+    int mPreviewheight;
     /**
      * Sets up member variables related to camera.
      *
      * @param width  The width of available size for camera preview
      * @param height The height of available size for camera preview
      */
-    int mPreviewwidth;
-    int mPreviewheight;
-
     private void setUpCameraOutputs(int width, int height) {
-        Activity activity = getActivity();
-        CameraManager manager = (CameraManager) activity.getSystemService(Context.CAMERA_SERVICE);
+        CameraManager manager = (CameraManager) Interface.i.mainActivity.getSystemService(Context.CAMERA_SERVICE);
         //CameraManager2 manager2 = new CameraManager2(manager);
         // manager2.CameraArr(manager);
         try {
-                CameraCharacteristics characteristics
-                        = manager.getCameraCharacteristics(Interface.i.settings.mCameraID);
-                mCameraCharacteristics = characteristics;
+            mCameraCharacteristics = manager.getCameraCharacteristics(Interface.i.settings.mCameraID);
                 mPreviewwidth = width;
                 mPreviewheight = height;
                 UpdateCameraCharacteristics(Interface.i.settings.mCameraID);
@@ -889,8 +799,7 @@ public class CameraFragment extends Fragment
     }
 
     public void UpdateCameraCharacteristics(String cameraId) {
-        Activity activity = getActivity();
-        CameraManager manager = (CameraManager) activity.getSystemService(Context.CAMERA_SERVICE);
+        CameraManager manager = (CameraManager) Interface.i.mainActivity.getSystemService(Context.CAMERA_SERVICE);
         CameraCharacteristics characteristics
                 = null;
         try {
@@ -901,8 +810,11 @@ public class CameraFragment extends Fragment
         mCameraCharacteristics = characteristics;
         //Integer facing = characteristics.get(CameraCharacteristics.LENS_FACING);
 
-        StreamConfigurationMap map = mCameraCharacteristics.get(
-                CameraCharacteristics.SCALER_STREAM_CONFIGURATION_MAP);
+        StreamConfigurationMap map = null;
+        if (mCameraCharacteristics != null) {
+            map = mCameraCharacteristics.get(
+                    CameraCharacteristics.SCALER_STREAM_CONFIGURATION_MAP);
+        }
         if (map == null) {
             return;
         }
@@ -926,22 +838,26 @@ public class CameraFragment extends Fragment
         Range[] ranges = characteristics.get(CameraCharacteristics.CONTROL_AE_AVAILABLE_TARGET_FPS_RANGES);
         int def = 30;
         int min = 20;
-        for(int i =0; i<ranges.length;i++){
-            if((int)ranges[i].getUpper() >= def){
-                FpsRangeDef = ranges[i];
-                break;
-            }
+        if(ranges == null) {
+            ranges = new Range[1];
+            ranges[0] = new Range(15, 30);
         }
-        if(FpsRangeDef == null)
-            for(int i =0; i<ranges.length;i++){
-                if((int)ranges[i].getUpper() >= min){
-                    FpsRangeDef = ranges[i];
+            for (Range value : ranges) {
+                if ((int) value.getUpper() >= def) {
+                    FpsRangeDef = value;
                     break;
                 }
             }
-        for(int i =0; i<ranges.length;i++){
-            if((int)ranges[i].getUpper() > def){
-                FpsRangeDef = ranges[i];
+        if(FpsRangeDef == null)
+            for (Range range : ranges) {
+                if ((int) range.getUpper() >= min) {
+                    FpsRangeDef = range;
+                    break;
+                }
+            }
+        for (Range range : ranges) {
+            if ((int) range.getUpper() > def) {
+                FpsRangeDef = range;
                 break;
             }
         }
@@ -966,7 +882,7 @@ public class CameraFragment extends Fragment
 
         mCameraAfModes = characteristics.get(CameraCharacteristics.CONTROL_AF_AVAILABLE_MODES);
         Point displaySize = new Point();
-        activity.getWindowManager().getDefaultDisplay().getSize(displaySize);
+        Interface.i.mainActivity.getWindowManager().getDefaultDisplay().getSize(displaySize);
         int rotatedPreviewWidth = mPreviewwidth;
         int rotatedPreviewHeight = mPreviewheight;
         int maxPreviewWidth = displaySize.x;
@@ -975,7 +891,9 @@ public class CameraFragment extends Fragment
         if (swappedDimensions) {
             rotatedPreviewWidth = mPreviewheight;
             rotatedPreviewHeight = mPreviewwidth;
+            //noinspection SuspiciousNameCombination
             maxPreviewWidth = displaySize.y;
+            //noinspection SuspiciousNameCombination
             maxPreviewHeight = displaySize.x;
         }
 
@@ -1007,7 +925,7 @@ public class CameraFragment extends Fragment
         // Check if the flash is supported.
         Boolean available = characteristics.get(CameraCharacteristics.FLASH_INFO_AVAILABLE);
         mFlashSupported = available == null ? false : available;
-        MainActivity.onCameraInitialization();
+        Interface.i.cameraui.onCameraInitialization();
     }
 
     @SuppressLint("MissingPermission")
@@ -1041,7 +959,11 @@ public class CameraFragment extends Fragment
         }
 
         Activity activity = getActivity();
-        CameraManager manager = (CameraManager) activity.getSystemService(Context.CAMERA_SERVICE);
+        CameraManager manager = null;
+        if (activity != null) {
+            manager = (CameraManager) activity.getSystemService(Context.CAMERA_SERVICE);
+        }
+        if(manager == null) return;
         StreamConfigurationMap map = null;
         try {
             map = manager.getCameraCharacteristics(Interface.i.settings.mCameraID).get(
@@ -1049,6 +971,7 @@ public class CameraFragment extends Fragment
         } catch (CameraAccessException e) {
             e.printStackTrace();
         }
+        if(map == null) return;
         Size preview = getCameraOutputSize(map.getOutputSizes(mPreviewTargetFormat));
         Size target = getCameraOutputSize(map.getOutputSizes(mTargetFormat),preview);
         int max = 3;
@@ -1087,19 +1010,18 @@ public class CameraFragment extends Fragment
     }*/
 
     /**
-     * Opens the camera specified by {@link CameraFragment#mCameraId}.
+     * Opens the camera specified by {@link CameraFragment#mCameraIds}.
      */
     private void openCamera(int width, int height) {
         context = this;
-        if (ContextCompat.checkSelfPermission(getActivity(), Manifest.permission.CAMERA)
+        if (ContextCompat.checkSelfPermission(Interface.i.mainActivity, Manifest.permission.CAMERA)
                 != PackageManager.PERMISSION_GRANTED) {
             //requestCameraPermission();
             return;
         }
         setUpCameraOutputs(width, height);
         configureTransform(width, height);
-        Activity activity = getActivity();
-        CameraManager manager = (CameraManager) activity.getSystemService(Context.CAMERA_SERVICE);
+        CameraManager manager = (CameraManager) Interface.i.mainActivity.getSystemService(Context.CAMERA_SERVICE);
         CameraManager2 manager2 = new CameraManager2(manager);
         mCameraIds = manager2.getCameraIdList();
         try {
@@ -1145,7 +1067,7 @@ public class CameraFragment extends Fragment
     /**
      * Starts a background thread and its {@link Handler}.
      */
-    private void startBackgroundThread() {
+    public void startBackgroundThread() {
         mBackgroundThread = new HandlerThread("CameraBackground");
         mBackgroundThread.start();
         mBackgroundHandler = new Handler(mBackgroundThread.getLooper());
@@ -1156,6 +1078,7 @@ public class CameraFragment extends Fragment
      * Stops the background thread and its {@link Handler}.
      */
     private void stopBackgroundThread() {
+        if(mBackgroundThread == null) return;
         mBackgroundThread.quitSafely();
         try {
             mBackgroundThread.join();
@@ -1188,7 +1111,7 @@ public class CameraFragment extends Fragment
             mPreviewRequestBuilder.addTarget(surface);
 
             // Here, we create a CameraCaptureSession for camera preview.
-            List surfaces = Arrays.asList(surface, mImageReaderPreview.getSurface());
+            List<Surface> surfaces = Arrays.asList(surface, mImageReaderPreview.getSurface());
             if(burst){
                 surfaces = Arrays.asList(mImageReaderPreview.getSurface(),mImageReaderRaw.getSurface());
             }
@@ -1209,15 +1132,17 @@ public class CameraFragment extends Fragment
                                     // Auto focus should be continuous for camera preview.
                                     //mPreviewRequestBuilder.set(CaptureRequest.CONTROL_AF_MODE,CaptureRequest.CONTROL_AF_MODE_CONTINUOUS_PICTURE);
                                     // Flash is automatically enabled when necessary.
-                                    setAutoFlash(mPreviewRequestBuilder);
+                                    setAutoFlash();
                                     Interface.i.settings.applyPrev(mPreviewRequestBuilder);
 
                                     //lightcycle.setVisibility(View.INVISIBLE);
                                     // Finally, we start displaying the camera preview.
                                     if (Interface.i.camera.is30Fps) {
-                                        mPreviewRequestBuilder.set(CaptureRequest.CONTROL_AE_TARGET_FPS_RANGE,FpsRangeDef);
+                                        mPreviewRequestBuilder.set(CaptureRequest.CONTROL_AE_TARGET_FPS_RANGE,
+                                                FpsRangeDef);
                                     } else {
-                                        mPreviewRequestBuilder.set(CaptureRequest.CONTROL_AE_TARGET_FPS_RANGE,FpsRangeHigh);
+                                        mPreviewRequestBuilder.set(CaptureRequest.CONTROL_AE_TARGET_FPS_RANGE,
+                                                FpsRangeHigh);
                                     }
                                     mPreviewRequest = mPreviewRequestBuilder.build();
 
@@ -1232,12 +1157,7 @@ public class CameraFragment extends Fragment
                                         burst = false;
                                     }
                                     if(getActivity()!=null){
-                                        getActivity().runOnUiThread(new Runnable() {
-                                            @Override
-                                            public void run() {
-                                                Interface.i.touchFocus.resetFocusCircle();
-                                            }
-                                        });
+                                        getActivity().runOnUiThread(() -> Interface.i.touchFocus.resetFocusCircle());
                                     }
                                 } catch (Exception e) {
                                     e.printStackTrace();
@@ -1368,11 +1288,6 @@ public class CameraFragment extends Fragment
         }
     }
 
-    /**
-     * Capture a still picture. This method should be called when we get a response in
-     * {@link #mCaptureCallback} from both {@link #lockFocus()}.
-     */
-    private boolean FixPreview = false;
     ArrayList<CaptureRequest> captures;
     CameraCaptureSession.CaptureCallback CaptureCallback;
     private void captureStillPicture() {
@@ -1385,7 +1300,6 @@ public class CameraFragment extends Fragment
             final CaptureRequest.Builder captureBuilder =
                     mCameraDevice.createCaptureRequest(CameraDevice.TEMPLATE_STILL_CAPTURE);
             mCaptureSession.stopRepeating();
-            FixPreview = true;
             if(mTargetFormat != mPreviewTargetFormat) captureBuilder.addTarget(mImageReaderRaw.getSurface());
             else captureBuilder.addTarget(mImageReaderPreview.getSurface());
             Interface.i.settings.applyRes(captureBuilder);
@@ -1404,10 +1318,11 @@ public class CameraFragment extends Fragment
             captureBuilder.set(CaptureRequest.JPEG_ORIENTATION, Interface.i.gravity.getCameraRotation());
             captures = new ArrayList<>();
             FrameNumberSelector.getFrames();
-            lightcycle.setMax(FrameNumberSelector.frameCount);
+            Interface.i.cameraui.lightcycle.setMax(FrameNumberSelector.frameCount);
             IsoExpoSelector.HDR = false;//Force HDR for tests
             captureBuilder.set(CaptureRequest.CONTROL_AF_MODE,CaptureRequest.CONTROL_AF_MODE_OFF);
             captureBuilder.set(CaptureRequest.LENS_FOCUS_DISTANCE,mFocus);
+            IsoExpoSelector.useTripod = Interface.i.sensors.getShakeness() < 5;
             for (int i = 0; i < FrameNumberSelector.frameCount; i++) {
                 IsoExpoSelector.setExpo(captureBuilder, i);
                 captures.add(captureBuilder.build());
@@ -1416,35 +1331,25 @@ public class CameraFragment extends Fragment
             Log.d(TAG,"FrameCount:"+FrameNumberSelector.frameCount);
             final int[] burstcount = {0, 0, FrameNumberSelector.frameCount};
             Log.d(TAG,"CaptureStarted!");
-            lightcycle.setAlpha(1.0f);
+            Interface.i.cameraui.lightcycle.setAlpha(1.0f);
             mTextureView.setAlpha(0.5f);
+            MediaPlayer burstPlayer = MediaPlayer.create(Interface.i.mainActivity,R.raw.sound_burst);
             CaptureCallback
                     = new CameraCaptureSession.CaptureCallback() {
                 @Override
                 public void onCaptureCompleted(@NonNull CameraCaptureSession session,
                                                @NonNull CaptureRequest request,
                                                @NonNull TotalCaptureResult result) {
-                    lightcycle.setProgress(lightcycle.getProgress() + 1);
-                    Log.d(TAG,"Completed!");
+                    Interface.i.cameraui.lightcycle.setProgress(Interface.i.cameraui.lightcycle.getProgress() + 1);
+                    burstPlayer.start();
+                    Log.v(TAG,"Completed!");
                     mCaptureResult = result;
-                    /*burstcount[0]++;
-                    if (burstcount[0] == burstcount[2] + 1 || ImageSaver.imageBuffer.size() == burstcount[2]) {
-                        try {
-                            mCaptureSession.abortCaptures();
-                            lightcycle.setAlpha(0f);
-                            lightcycle.setProgress(0);
-                            mTextureView.setAlpha(1f);
-                            unlockFocus();
-                        } catch (CameraAccessException e) {
-                            e.printStackTrace();
-                        }
-                    }*/
-
                 }
 
                 @Override
                 public void onCaptureStarted(@NonNull CameraCaptureSession session, @NonNull CaptureRequest request, long timestamp, long frameNumber) {
-                    Log.d(TAG,"FrameCaptureStarted! FrameNumber:"+frameNumber);
+                    burstPlayer.seekTo(0);
+                    Log.v(TAG,"FrameCaptureStarted! FrameNumber:"+frameNumber);
                     super.onCaptureStarted(session, request, timestamp, frameNumber);
                 }
 
@@ -1452,15 +1357,14 @@ public class CameraFragment extends Fragment
                 public void onCaptureSequenceCompleted(@NonNull CameraCaptureSession session, int sequenceId, long frameNumber) {
                     Log.d(TAG,"SequenceCompleted");
                     try {
-                        lightcycle.setAlpha(0f);
-                        lightcycle.setProgress(0);
+                        Interface.i.cameraui.lightcycle.setAlpha(0f);
+                        Interface.i.cameraui.lightcycle.setProgress(0);
                         mTextureView.setAlpha(1f);
                     } catch (Exception e){
                     e.printStackTrace();
                     }
                     //unlockFocus();
                     createCameraPreviewSession();
-                    FixPreview = false;
                     super.onCaptureSequenceCompleted(session, sequenceId, frameNumber);
                 }
 
@@ -1470,11 +1374,10 @@ public class CameraFragment extends Fragment
                     if (burstcount[1] >= burstcount[2] + 1 || ImageSaver.imageBuffer.size() >= burstcount[2]) {
                         try {
                             mCaptureSession.abortCaptures();
-                            lightcycle.setAlpha(0f);
-                            lightcycle.setProgress(0);
+                            Interface.i.cameraui.lightcycle.setAlpha(0f);
+                            Interface.i.cameraui.lightcycle.setProgress(0);
                             mTextureView.setAlpha(1f);
                             createCameraPreviewSession();
-                            FixPreview = false;
                         } catch (CameraAccessException e) {
                             e.printStackTrace();
                         }
@@ -1493,20 +1396,6 @@ public class CameraFragment extends Fragment
     }
 
     /**
-     * Retrieves the JPEG orientation from the specified screen rotation.
-     *
-     * @param rotation The screen rotation.
-     * @return The JPEG orientation (one of 0, 90, 270, and 360)
-     */
-    public int getOrientation(int rotation) {
-        // Sensor orientation is 90 for most devices, or 270 for some devices (eg. Nexus 5X)
-        // We have to take that into account and rotate JPEG properly.
-        // For devices with orientation of 90, we simply return our mapping from ORIENTATIONS.
-        // For devices with orientation of 270, we need to rotate the JPEG 180 degrees.
-        return (ORIENTATIONS.get(rotation) + mSensorOrientation + 270) % 360;
-    }
-
-    /**
      * Unlock the focus. This method should be called when still image capture sequence is
      * finished.
      */
@@ -1514,7 +1403,7 @@ public class CameraFragment extends Fragment
         // Reset the auto-focus trigger
         //mCaptureSession.stopRepeating();
         CameraReflectionApi.set(mPreviewRequest,CaptureRequest.CONTROL_AF_TRIGGER,CameraMetadata.CONTROL_AF_TRIGGER_CANCEL);
-        setAutoFlash(mPreviewRequestBuilder);
+        setAutoFlash();
         //mCaptureSession.capture(mPreviewRequest, mCaptureCallback,
         //        mBackgroundHandler);
         // After this, the camera will go back to the normal state of preview.
@@ -1536,16 +1425,12 @@ public class CameraFragment extends Fragment
         return ids[n];
     }
 
-    public void setAutoFlash(CaptureRequest.Builder requestBuilder) {
+    public void setAutoFlash() {
         if (mFlashSupported) {
             if (mFlashEnabled)
             CameraReflectionApi.set(mPreviewRequest,CaptureRequest.CONTROL_AE_MODE,CaptureRequest.CONTROL_AE_MODE_ON_AUTO_FLASH);
         }
     }
-
-    /**
-     * Saves a JPEG {@link Image} into the specified {@link File}.
-     */
 
 
     /**
@@ -1580,11 +1465,11 @@ public class CameraFragment extends Fragment
         @Override
         public Dialog onCreateDialog(Bundle savedInstanceState) {
             final Activity activity = getActivity();
+            assert getArguments() != null;
             return new AlertDialog.Builder(activity)
                     .setMessage(getArguments().getString(ARG_MESSAGE))
-                    .setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialogInterface, int i) {
+                    .setPositiveButton(android.R.string.ok, (dialogInterface, i) -> {
+                        if (activity != null) {
                             activity.finish();
                         }
                     })
@@ -1595,7 +1480,7 @@ public class CameraFragment extends Fragment
     /**
      * Start the timer for the pre-capture sequence.
      * <p/>
-     * Call this only with {@link #mCameraStateLock} held.
+     * Call this only with { #mCameraStateLock} held.
      */
     private void startTimerLocked() {
         mCaptureTimer = SystemClock.elapsedRealtime();
@@ -1604,7 +1489,7 @@ public class CameraFragment extends Fragment
     /**
      * Check if the timer for the pre-capture sequence has been hit.
      * <p/>
-     * Call this only with {@link #mCameraStateLock} held.
+     * Call this only with { #mCameraStateLock} held.
      *
      * @return true if the timeout occurred.
      */
