@@ -1,5 +1,6 @@
 package com.eszdman.photoncamera.Control;
 
+import android.annotation.SuppressLint;
 import android.graphics.Point;
 import android.graphics.Rect;
 import android.hardware.camera2.CameraCharacteristics;
@@ -17,11 +18,12 @@ import com.eszdman.photoncamera.api.Interface;
 import com.eszdman.photoncamera.ui.CameraFragment;
 
 public class TouchFocus {
-    private final String TAG = "TouchFocus";
+    protected final String TAG = "TouchFocus";
     boolean activated = false;
     public boolean onConfigured = false;
     ImageView focusEl;
     AutoFitTextureView preview;
+    @SuppressLint("ClickableViewAccessibility")
     public void ReInit(){
         focusEl = Interface.i.mainActivity.findViewById(R.id.touchFocus);
         focusEl.setOnTouchListener(focusListener);
@@ -31,6 +33,7 @@ public class TouchFocus {
     final OnTouchListener focusListener = new OnTouchListener() {
         @Override
         public boolean onTouch(View v, MotionEvent event) {
+            v.performClick();
             activated = false;
             focusEl.setVisibility(View.GONE);
             focusEl.setX((float) getMax().x/2.f);
@@ -47,7 +50,7 @@ public class TouchFocus {
         }
     };
 
-    public void processTochToFocus(View v, float fx, float fy) {
+    public void processTochToFocus(float fx, float fy) {
         activated = true;
         focusEl.setX(fx-focusEl.getMeasuredWidth()/2.0f);
         focusEl.setY(fy-focusEl.getMeasuredHeight()/2.0f);
@@ -74,24 +77,25 @@ public class TouchFocus {
             x = 0;
         if (y<0)
             y = 0;
-        if (y > CurUi.y)
+        /*if (y > CurUi.y)
             y = CurUi.y;
         if (x > CurUi.x)
-            x  =CurUi.x;
+            x  =CurUi.x;*/
         //use 1/8 from the the sensor size for the focus rect
-        int width_to_set = sizee.width()/8;
-        int height_to_set = width_to_set;
-        float x_scale = (float)sizee.width()/(float)CurUi.x;
-        float y_scale = (float)sizee.height()/(float)CurUi.y;
-        int x_to_set = (int)(x * x_scale) -width_to_set/2;
+        int width_to_set = sizee.width()/6;
+        float kprop = (float)CurUi.x/(float)(CurUi.y);
+        int height_to_set = (int)(width_to_set*kprop);
+        float x_scale = (float)sizee.width()/(float)CurUi.y;
+        float y_scale = (float)sizee.height()/(float)CurUi.x;
+        int x_to_set = (int)(x * x_scale) - width_to_set/2;
         int y_to_set = (int)(y * y_scale) - height_to_set/2;
         if (x_to_set < 0)
             x_to_set = 0;
         if (y_to_set < 0)
             y_to_set = 0;
-        if (y_to_set + height_to_set/2 > sizee.height())
+        if (y_to_set - height_to_set > sizee.height())
             y_to_set = sizee.height() - height_to_set;
-        if (x_to_set + width_to_set/2 > sizee.width())
+        if (x_to_set - width_to_set  > sizee.width())
             y_to_set = sizee.width() - width_to_set;
         MeteringRectangle rect_to_set =new MeteringRectangle(x_to_set,y_to_set, width_to_set,height_to_set,MeteringRectangle.METERING_WEIGHT_MAX-1);
         MeteringRectangle[] rectaf = new MeteringRectangle[1];
@@ -122,9 +126,6 @@ public class TouchFocus {
     }
     public Point getMax(){
         return new Point(preview.getWidth(),preview.getHeight());
-    }
-    public Point getCurrent(){
-        return new Point((int)(focusEl.getX()+150.f),(int)(focusEl.getY()-110.f));
     }
     public void resetFocusCircle(){ //resets the position and visibility of focus circle
         focusEl.setVisibility(View.GONE);
