@@ -16,14 +16,14 @@ public class GLInterface {
         glProc = processing;
         glprogram = glProc.mProgram;
     }
+
     public GLInterface(GLContext context){
         glContext = context;
         glprogram = glContext.mProgram;
     }
-    @SuppressLint("NewApi")
+
     static public String loadShader(int fragment) {
         StringBuilder source = new StringBuilder();
-
         BufferedReader reader = new BufferedReader(new InputStreamReader(Interface.i.mainActivity.getResources().openRawResource(fragment)));
         for (Object line : reader.lines().toArray()) {
             source.append(line).append("\n");
