@@ -19,10 +19,9 @@ public class GLOneParams {
         glProc = glCoreBlockProcessing;
         glprogram = glProc.mProgram;
     }
-    @SuppressLint("NewApi")
+
     static public String loadShader(int fragment) {
         StringBuilder source = new StringBuilder();
-
         BufferedReader reader = new BufferedReader(new InputStreamReader(Interface.i.mainActivity.getResources().openRawResource(fragment)));
         for (Object line : reader.lines().toArray()) {
             source.append(line).append("\n");
