@@ -16,8 +16,9 @@ import java.lang.reflect.Field;
 
 import static android.hardware.camera2.CaptureResult.*;
 import static android.hardware.camera2.CameraCharacteristics.*;
+@SuppressWarnings("ALL")
 public class Camera2ApiAutoFix {
-    private static String TAG = "Camera2ApiAutoFix";
+    private static final String TAG = "Camera2ApiAutoFix";
     private CameraCharacteristics characteristics;
     private CaptureResult result;
     Camera2ApiAutoFix(CameraCharacteristics characteristic) {
@@ -39,11 +40,15 @@ public class Camera2ApiAutoFix {
         CaptureResult characteristics= CameraFragment.mCaptureResult;
         Camera2ApiAutoFix fix = new Camera2ApiAutoFix(characteristics);
         //fix.gains();
-        fix.dynBL();
+        fix.BL();
         fix.whitePoint();
+        fix.CCM();
     }
     private void whitePoint(){
         CameraReflectionApi.set(SENSOR_NEUTRAL_COLOR_POINT,Interface.i.camera.mPreviewTemp);
+    }
+    private void CCM(){
+        CameraReflectionApi.set(COLOR_CORRECTION_TRANSFORM,Interface.i.camera.mColorSpaceTransform);
     }
     public void curve(){
         CameraReflectionApi.set(TONEMAP_MAX_CURVE_POINTS,128);
@@ -98,7 +103,7 @@ public class Camera2ApiAutoFix {
         Log.d(TAG,"Overrided channelVector:"+rggbChannelVector.toString());
     }
     @SuppressLint("NewApi")
-    public void dynBL(){
+    public void BL(){
        float[] level = result.get(SENSOR_DYNAMIC_BLACK_LEVEL);
         BlackLevelPattern ptr = CameraFragment.mCameraCharacteristics.get(CameraCharacteristics.SENSOR_BLACK_LEVEL_PATTERN);
         if(ptr == null) return;
