@@ -5,7 +5,6 @@ import android.hardware.camera2.CameraMetadata;
 import android.hardware.camera2.CaptureRequest;
 import android.hardware.camera2.CaptureResult;
 import android.hardware.camera2.params.BlackLevelPattern;
-import android.util.Log;
 
 import com.eszdman.photoncamera.ui.CameraFragment;
 
@@ -13,17 +12,17 @@ import org.chickenhook.restrictionbypass.RestrictionBypass;
 
 import java.lang.reflect.Array;
 import java.lang.reflect.Field;
-import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 
 public class CameraReflectionApi {
-    private static final String TAG = "CameraReflectionApi";
+    //private static final String TAG = "CameraReflectionApi";
 
     public static <T> void set(CameraCharacteristics.Key<T> key, T value) {
         try {
             Field CameraMetadataNativeField = RestrictionBypass.getDeclaredField(CameraCharacteristics.class, "mProperties");
             CameraMetadataNativeField.setAccessible(true);
             Object CameraMetadataNative = CameraMetadataNativeField.get(CameraFragment.mCameraCharacteristics);//Ur camera Characteristics
+            assert CameraMetadataNative != null;
             Method set = RestrictionBypass.getDeclaredMethod(CameraMetadataNative.getClass(),"set",CameraCharacteristics.Key.class, Object.class);
             set.setAccessible(true);
             set.invoke(CameraMetadataNative, key, value);
@@ -37,6 +36,7 @@ public class CameraReflectionApi {
             Field CameraMetadataNativeField = RestrictionBypass.getDeclaredField(CaptureResult.class,"mResults");
             CameraMetadataNativeField.setAccessible(true);
             Object CameraMetadataNative = CameraMetadataNativeField.get(CameraFragment.mCaptureResult);
+            assert CameraMetadataNative != null;
             Method set = RestrictionBypass.getDeclaredMethod(CameraMetadataNative.getClass(),"set", CaptureResult.Key.class, Object.class);
             set.setAccessible(true);
             set.invoke(CameraMetadataNative, key, value);
@@ -49,6 +49,7 @@ public class CameraReflectionApi {
             Field CameraMetadataNativeField = RestrictionBypass.getDeclaredField(CaptureResult.class,"mResults");
             CameraMetadataNativeField.setAccessible(true);
             Object CameraMetadataNative = CameraMetadataNativeField.get(res);
+            assert CameraMetadataNative != null;
             Method set = RestrictionBypass.getDeclaredMethod(CameraMetadataNative.getClass(),"set", CaptureResult.Key.class, Object.class);
             set.setAccessible(true);
             set.invoke(CameraMetadataNative, key, value);
@@ -61,6 +62,7 @@ public class CameraReflectionApi {
             Field CameraMetadataNativeField = RestrictionBypass.getDeclaredField(CaptureRequest.class,"mLogicalCameraSettings");
             CameraMetadataNativeField.setAccessible(true);
             Object CameraMetadataNative = CameraMetadataNativeField.get(request);
+            assert CameraMetadataNative != null;
             Method set = RestrictionBypass.getDeclaredMethod(CameraMetadataNative.getClass(),"set", CaptureRequest.Key.class, Object.class);
             set.setAccessible(true);
             set.invoke(CameraMetadataNative, key, value);
@@ -70,10 +72,12 @@ public class CameraReflectionApi {
     }
     public static void PatchBL(BlackLevelPattern pattern, int[] bl){
         try {
+            //noinspection JavaReflectionMemberAccess
             Field mCfaOffsetsField = pattern.getClass().getDeclaredField("mCfaOffsets");
             mCfaOffsetsField.setAccessible(true);
             Object mCfaOffsets = mCfaOffsetsField.get(pattern);
             for(int i =0; i<4;i++){
+                assert mCfaOffsets != null;
                 Array.set(mCfaOffsets,i,bl[i]);
             }
         } catch (NoSuchFieldException | IllegalAccessException e) {
@@ -81,55 +85,27 @@ public class CameraReflectionApi {
         }
 
     }
-    public static void PrintFields(Object in){
-        Log.d(TAG,"StartPrinting:"+in.getClass());
-        Field[] fields = in.getClass().getDeclaredFields();
-        int cnt = 0;
-        for(Field f : fields) {Log.d(TAG,"["+cnt+"]"+f.toString());cnt++;}
-    }
+
     public static Field[] getAllMetadataFields(){
         return  CameraMetadata.class.getDeclaredFields();
     }
-    public static void PrintMethods(Object in){
+    /*public static void PrintMethods(Object in){
         Log.d(TAG,"StartPrinting:"+in.getClass());
         Method[] methods = in.getClass().getDeclaredMethods();
         int cnt = 0;
         for(Method m : methods) {Log.d(TAG,"["+cnt+"]"+m.toString());cnt++;}
     }
-    public static void native_set(String key, String val){
-        try {
-            Class SystemProperties = Class.forName("android.os.SystemProperties");
-            Method set = RestrictionBypass.getDeclaredMethod(SystemProperties,"set",String.class,String.class);
-            set.setAccessible(true);
-            set.invoke(null,key,val);
-        } catch (ClassNotFoundException | NoSuchMethodException | IllegalAccessException | InvocationTargetException e) {
-            e.printStackTrace();
-        }
-
-    }
-    public static void setAuxOn(){
-        Log.d(TAG,"Setting Aux ON");
-        try {
-            /*Class actThread = Class.forName("android.app.ActivityThread");
-            Method getCur = RestrictionBypass.getDeclaredMethod(actThread,"currentActivityThread");
-            getCur.setAccessible(true);
-            Object curThread = getCur.invoke(null);
-            Field isSystem = RestrictionBypass.getDeclaredField(curThread.getClass(),"mSystemThread");
-            isSystem.set(curThread,true);*/
-            Class SystemProperties = Class.forName("android.os.SystemProperties");
-            Method set = RestrictionBypass.getDeclaredMethod(SystemProperties,"native_set",String.class,String.class);
-            set.setAccessible(true);
-            //set.invoke(null,"camera.aux.packagelist", "com.eszdman.photoncamera");//ActivityThread.currentPackageName()
-        } catch (Exception e){
-            e.printStackTrace();
-        }
-        Log.d(TAG,"Setting Aux done!");
-    }
-
+    public static void PrintFields(Object in){
+        Log.d(TAG,"StartPrinting:"+in.getClass());
+        Field[] fields = in.getClass().getDeclaredFields();
+        int cnt = 0;
+        for(Field f : fields) {Log.d(TAG,"["+cnt+"]"+f.toString());cnt++;}
+    }*/
     public static void setVERBOSE(boolean in) {
         Object capres = CameraFragment.mCaptureResult;//Ur camera CaptureResult
-        Field verbose = null;
+        Field verbose;
         try {
+            //noinspection JavaReflectionMemberAccess
             verbose =  CaptureResult.class.getDeclaredField("VERBOSE");
             verbose.setAccessible(true);
             verbose.set(capres, in);
