@@ -3,42 +3,30 @@ package com.eszdman.photoncamera.api;
 import android.content.Intent;
 import android.database.Cursor;
 import android.graphics.Bitmap;
-import android.media.ThumbnailUtils;
 import android.net.Uri;
-import android.os.Build;
 import android.os.Handler;
 import android.os.Message;
 import android.provider.MediaStore;
 import android.util.Log;
-import android.util.Size;
 
 import com.eszdman.photoncamera.ui.MainActivity;
 
 import java.io.File;
 import java.util.Comparator;
 
+import rapid.decoder.BitmapDecoder;
+
+@SuppressWarnings("ALL")
 public class Photo {
     public static Photo instance;
     private static Handler galleryHandler;
     static class GalleryHandler extends Handler {
-        GalleryHandler() {
-        }
         @Override
         public void handleMessage(Message msg)
         {
-            Uri uri = null;
-            Bitmap bmp = null;
-            if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-                bmp = (Bitmap) msg.obj;
-            } else{
-                uri = (Uri)msg.obj;
-            }
+            Bitmap bmp = (Bitmap) msg.obj;
             try {
-                if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-                    Interface.i.camera.galleryImageButton.setImageBitmap(bmp);
-                } else {
-                    Interface.i.camera.galleryImageButton.setImageURI(uri);
-                }
+              Interface.i.cameraui.galleryImageButton.setImageBitmap(bmp);
             } catch (Exception e) {
                 e.printStackTrace();
             }
@@ -99,12 +87,8 @@ public class Photo {
         try {
             Bitmap thumb = null;
             Message uriMessage = new Message();
-            if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-                thumb = ThumbnailUtils.createImageThumbnail(in, new Size(120, 120), null);
-                uriMessage.obj = thumb;
-            } else {
-                uriMessage.obj = contentUri;
-            }
+            thumb = BitmapDecoder.from(Uri.fromFile(in)).scaleBy(0.1f).decode();
+            uriMessage.obj = thumb;
             galleryHandler.sendMessage(uriMessage);
         } catch (Exception e) {
             e.printStackTrace();
