@@ -2,17 +2,22 @@ package com.eszdman.photoncamera.api;
 
 import android.annotation.SuppressLint;
 import android.content.SharedPreferences;
-import android.graphics.Point;
 import android.hardware.camera2.CaptureRequest;
 import android.hardware.camera2.params.MeteringRectangle;
 import android.hardware.camera2.params.TonemapCurve;
+import android.os.Environment;
 import android.util.Log;
 import android.util.Range;
 
 import com.eszdman.photoncamera.ui.CameraFragment;
 import com.eszdman.photoncamera.ui.MainActivity;
 
-import java.net.InterfaceAddress;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.util.Map;
+import java.util.Properties;
 
 import static android.content.Context.MODE_PRIVATE;
 import static android.hardware.camera2.CameraCharacteristics.CONTROL_AE_COMPENSATION_RANGE;
@@ -20,7 +25,6 @@ import static android.hardware.camera2.CameraMetadata.CONTROL_AE_MODE_ON;
 import static android.hardware.camera2.CameraMetadata.CONTROL_AE_STATE_LOCKED;
 import static android.hardware.camera2.CameraMetadata.CONTROL_AF_MODE_CONTINUOUS_PICTURE;
 import static android.hardware.camera2.CameraMetadata.HOT_PIXEL_MODE_HIGH_QUALITY;
-import static android.hardware.camera2.CameraMetadata.LENS_OPTICAL_STABILIZATION_MODE_OFF;
 import static android.hardware.camera2.CameraMetadata.LENS_OPTICAL_STABILIZATION_MODE_ON;
 import static android.hardware.camera2.CameraMetadata.NOISE_REDUCTION_MODE_HIGH_QUALITY;
 import static android.hardware.camera2.CameraMetadata.NOISE_REDUCTION_MODE_OFF;
@@ -36,6 +40,7 @@ import static android.hardware.camera2.CaptureRequest.NOISE_REDUCTION_MODE;
 import static android.hardware.camera2.CaptureRequest.TONEMAP_CURVE;
 import static android.hardware.camera2.CaptureRequest.TONEMAP_MODE;
 
+@SuppressWarnings("ALL")
 public class Settings {
     private final String TAG = "Settings";
     public int noiseReduction = NOISE_REDUCTION_MODE_OFF;
@@ -51,11 +56,12 @@ public class Settings {
     public boolean enhancedProcess = false;
     public boolean grid = false;
     public boolean watermark = false;
+    public boolean afdata = false;
     public boolean roundedge = true;
     public boolean align = true;
     public boolean hdrx = true;
     public boolean hdrxNR = true;
-    public double saturation = 1.2;
+    public double saturation = 1.0;
     public double sharpness = 0.7;
     public double contrastMpy = 1.0;
     public int contrastConst = 0;
@@ -69,10 +75,19 @@ public class Settings {
     public boolean remosaic = false;
     public boolean eisPhoto = true;
     public boolean fpsPreview = false;
-    public boolean nightMode = false;
     public int alignAlgorithm = 0;
     public String mCameraID = "0";
     private int count = 0;
+    public int selectedMode = CameraMode.DEFAULT.mNum;
+    public enum CameraMode {
+        DEFAULT(0),
+        NIGHT(1),
+        UNLIMITED(2);
+        public int mNum;
+        CameraMode(int number) {
+            mNum = number;
+        }
+    }
     private SharedPreferences.Editor sharedPreferencesEditor;
     private SharedPreferences sharedPreferences;
     Settings(){
@@ -87,118 +102,68 @@ public class Settings {
     }
     public void load() {
         sharedPreferences = MainActivity.act.getPreferences(MODE_PRIVATE);
-        noiseReduction = get(noiseReduction);
-        Log.d(TAG, "Loaded noise reduction:" + noiseReduction);
-        frameCount = get(frameCount);
-        Log.d(TAG, "Loaded frame count:" + frameCount);
-        align = get(align);
-        Log.d(TAG, "Loaded align:" + align);
-        lumenCount = get(lumenCount);
-        Log.d(TAG, "Loaded lumen count:" + lumenCount);
-        chromaCount = get(chromaCount);
-        Log.d(TAG, "Loaded chroma count:" + chromaCount);
-        enhancedProcess = get(enhancedProcess);
-        Log.d(TAG, "Loaded enhanced process:" + enhancedProcess);
-        grid = get(grid);
-        Log.d(TAG, "Loaded grid:" + grid);
-        watermark = get(watermark);
-        Log.d(TAG, "Loaded watermark:" + watermark);
-        roundedge = get(roundedge);
-        Log.d(TAG, "Loaded round edges:" + roundedge);
-        sharpness = get(sharpness);
-        Log.d(TAG, "Loaded sharpness:" + sharpness);
-        contrastMpy = get(contrastMpy);
-        Log.d(TAG, "Loaded contrast mpy:" + contrastMpy);
-        contrastConst = get(contrastConst);
-        Log.d(TAG, "Loaded contrast const:" + contrastConst);
-        saturation = get(saturation);
-        Log.d(TAG, "Loaded saturation:" + saturation);
-        compressor = get(compressor);
-        Log.d(TAG, "Loaded compressor:" + compressor);
-        gain = get(gain);
-        Log.d(TAG, "Loaded gain:" + gain);
-        lastPicture = get(lastPicture);
-        Log.d(TAG, "Loaded last picture:" + lastPicture);
-        hdrx = get(hdrx);
-        Log.d(TAG, "Loaded hdrx:" + hdrx);
-        cfaPattern = get(cfaPattern);
-        Log.d(TAG, "Loaded CFA:" + cfaPattern);
-        rawSaver = get(rawSaver);
-        Log.d(TAG, "Loaded rawSaver:" + rawSaver);
-        remosaic = get(remosaic);
-        Log.d(TAG, "Loaded remosaic:" + remosaic);
-        eisPhoto = get(eisPhoto);
-        Log.d(TAG, "Loaded eisPhoto:" + eisPhoto);
-        QuadBayer = get(QuadBayer);
-        Log.d(TAG, "Loaded QuadBayer:" + QuadBayer);
-        fpsPreview = get(fpsPreview);
-        Log.d(TAG, "Loaded fpsPreview:" + fpsPreview);
-        hdrxNR = get(hdrxNR);
-        Log.d(TAG,"Loaded hdrxNR:"+hdrxNR);
-        alignAlgorithm = get(alignAlgorithm);
-
+        noiseReduction = get(noiseReduction,"NoiseReduction");
+        frameCount = get(frameCount,"FrameCount");
+        align = get(align,"Alignment");
+        lumenCount = get(lumenCount,"LumaCount");
+        chromaCount = get(chromaCount,"ChromaCount");
+        enhancedProcess = get(enhancedProcess,"EnhancedProc");
+        grid = get(grid,"Grid");
+        watermark = get(watermark,"Watermark");
+        afdata = get(afdata,"AFData");
+        roundedge = get(roundedge,"RoundEdges");
+        sharpness = get(sharpness,"Sharpness");
+        contrastMpy = get(contrastMpy,"ContrastMpy");
+        contrastConst = get(contrastConst,"ContrastConst");
+        saturation = get(saturation,"Saturation");
+        compressor = get(compressor,"Compressor");
+        gain = get(gain,"Gain");
+        hdrx = get(hdrx,"Hdrx");
+        cfaPattern = get(cfaPattern,"CFA");
+        rawSaver = get(rawSaver,"SaveRaw");
+        remosaic = get(remosaic,"Remosaic");
+        eisPhoto = get(eisPhoto,"EisPhoto");
+        QuadBayer = get(QuadBayer,"QuadBayer");
+        fpsPreview = get(fpsPreview,"FpsPreview");
+        hdrxNR = get(hdrxNR,"HdrxNR");
+        alignAlgorithm = get(alignAlgorithm,"AlignmentAlgo");
+        selectedMode = get(selectedMode, "SelectedMode");
         count = -1;
-        mCameraID = get(mCameraID);
-        Log.d(TAG, "Loaded mCameraID:" + mCameraID);
+        mCameraID = sharedPreferences.getString("Camera", mCameraID);
         count = 0;
     }
 
     public void save() {
         sharedPreferences = MainActivity.act.getPreferences(MODE_PRIVATE);
         sharedPreferencesEditor = sharedPreferences.edit();
-        put(noiseReduction);
-        Log.d(TAG, "Saved noise reduction:" + noiseReduction);
-        put(frameCount);
-        Log.d(TAG, "Saved frame count:" + frameCount);
-        put(align);
-        Log.d(TAG, "Saved align:" + align);
-        put(lumenCount);
-        Log.d(TAG, "Saved lumen count:" + lumenCount);
-        put(chromaCount);
-        Log.d(TAG, "Saved chroma count:" + chromaCount);
-        put(enhancedProcess);
-        Log.d(TAG, "Saved enhanced process:" + enhancedProcess);
-        put(grid);
-        Log.d(TAG, "Saved grid:" + grid);
-        put(watermark);
-        Log.d(TAG, "Saved watermark:" + watermark);
-        put(roundedge);
-        Log.d(TAG, "Saved round edges:" + roundedge);
-        put(sharpness);
-        Log.d(TAG, "Saved sharpness:" + sharpness);
-        put(contrastMpy);
-        Log.d(TAG, "Saved contrast mpy:" + contrastMpy);
-        put(contrastConst);
-        Log.d(TAG, "Saved contrast const:" + contrastConst);
-        put(saturation);
-        Log.d(TAG, "Saved saturation:" + saturation);
-        put(compressor);
-        Log.d(TAG, "Saved compressor:" + compressor);
-        put(gain);
-        Log.d(TAG, "Saved gain:" + gain);
-        put(lastPicture);
-        Log.d(TAG, "Saved last picture:" + lastPicture);
-        put(hdrx);
-        Log.d(TAG, "Saved hdrx:" + hdrx);
-        Log.d(TAG, "Saved CFA:" + cfaPattern);
-        put(cfaPattern);
-        Log.d(TAG, "Saved RawSaver:" + rawSaver);
-        put(rawSaver);
-        Log.d(TAG, "Saved remosaic:" + remosaic);
-        put(remosaic);
-        Log.d(TAG, "Saved eisPhoto:" + eisPhoto);
-        put(eisPhoto);
-        Log.d(TAG, "Saved QuadBayer:" + QuadBayer);
-        put(QuadBayer);
-        Log.d(TAG, "Saved fpsPreview:" + fpsPreview);
-        put(fpsPreview);
-        Log.d(TAG,"Saved hdrxNR:"+hdrxNR);
-        put(hdrxNR);
-        put(alignAlgorithm);
+        put(noiseReduction,"NoiseReduction");
+        put(frameCount,"FrameCount");
+        put(align,"Alignment");
+        put(lumenCount,"LumaCount");
+        put(chromaCount,"ChromaCount");
+        put(enhancedProcess,"EnhancedProc");
+        put(grid,"Grid");
+        put(watermark,"Watermark");
+        put(afdata,"AFData");
+        put(roundedge,"RoundEdges");
+        put(sharpness,"Sharpness");
+        put(contrastMpy,"ContrastMpy");
+        put(contrastConst,"ContrastConst");
+        put(saturation,"Saturation");
+        put(compressor,"Compressor");
+        put(gain,"Gain");
+        put(hdrx,"Hdrx");
+        put(cfaPattern,"CFA");
+        put(rawSaver,"SaveRaw");
+        put(remosaic,"Remosaic");
+        put(eisPhoto,"EisPhoto");
+        put(QuadBayer,"QuadBayer");
+        put(fpsPreview,"FpsPreview");
+        put(hdrxNR,"HdrxNR");
+        put(alignAlgorithm,"AlignmentAlgo");
+        put(selectedMode, "SelectedMode");
         count = -1;
-        Log.d(TAG, "Saved mCameraID:" + mCameraID);
-        put(mCameraID);
-
+        sharedPreferencesEditor.putString("Camera", mCameraID);
         sharedPreferencesEditor.apply();
         count = 0;
     }
@@ -207,14 +172,12 @@ public class Settings {
         count = -1;
         sharedPreferences = MainActivity.act.getPreferences(MODE_PRIVATE);
         sharedPreferencesEditor = sharedPreferences.edit();
-        Log.d(TAG, "Saved mCameraID:" + mCameraID);
-        put(mCameraID);
+        sharedPreferencesEditor.putString("Camera", mCameraID);
         sharedPreferencesEditor.apply();
         count = 0;
     }
     public void applyRes(CaptureRequest.Builder captureBuilder) {
         captureBuilder.set(HOT_PIXEL_MODE, HOT_PIXEL_MODE_HIGH_QUALITY);
-        //captureBuilder.set(COLOR_CORRECTION_MODE, COLOR_CORRECTION_MODE_HIGH_QUALITY);
         captureBuilder.set(CONTROL_AE_MODE, CaptureRequest.CONTROL_AE_MODE_OFF);
         //captureBuilder.set(CONTROL_AF_MODE, CONTROL_AF_MODE_OFF);
         //captureBuilder.set(STATISTICS_LENS_SHADING_MAP_MODE, STATISTICS_LENS_SHADING_MAP_MODE_ON);
@@ -226,14 +189,14 @@ public class Settings {
         Camera2ApiAutoFix.Apply();
         captureBuilder.set(NOISE_REDUCTION_MODE, NOISE_REDUCTION_MODE_HIGH_QUALITY);
         captureBuilder.set(CONTROL_AE_MODE, aeModeOn);
+        //captureBuilder.set(COLOR_CORRECTION_MODE,COLOR_CORRECTION_MODE_HIGH_QUALITY);
         captureBuilder.set(LENS_OPTICAL_STABILIZATION_MODE,LENS_OPTICAL_STABILIZATION_MODE_ON);//Fix ois bugs for preview and burst
         //captureBuilder.set(CONTROL_AE_EXPOSURE_COMPENSATION,-1);
         Range range = CameraFragment.mCameraCharacteristics.get(CONTROL_AE_COMPENSATION_RANGE);
-
-        if(nightMode) captureBuilder.set(CONTROL_AE_EXPOSURE_COMPENSATION,(int)range.getUpper());
-        Point size = new Point(Interface.i.camera.mImageReaderPreview.getWidth(),Interface.i.camera.mImageReaderPreview.getHeight());
+        if(selectedMode == CameraMode.NIGHT.mNum && range != null) captureBuilder.set(CONTROL_AE_EXPOSURE_COMPENSATION,(int)range.getUpper());
+        /*Point size = new Point(Interface.i.camera.mImageReaderPreview.getWidth(),Interface.i.camera.mImageReaderPreview.getHeight());
         double sizex = size.x;
-        double sizey = size.y;
+        double sizey = size.y;*/
         //captureBuilder.set(CONTROL_AE_TARGET_FPS_RANGE,new Range<>(24,60));
         /*MeteringRectangle[] rectm8 = new MeteringRectangle[2];
         rectm8[0] = new MeteringRectangle(new Point((int)(sizex/2.0),(int)(sizey/2.0)),new Size((int)(sizex*2.0/4.0),(int)(sizey*2.0/4.0)),10);
@@ -264,56 +227,130 @@ public class Settings {
         }
         TonemapCurve tonemapCurve = new TonemapCurve(rgb,rgb,rgb);
         captureBuilder.set(TONEMAP_CURVE,tonemapCurve);
-        //final CaptureRequest.Key<Boolean> EISV2 = new CaptureRequest.Key<Boolean>("com.qti.node.eisv2", boolean.class);
-        //captureBuilder.set(EISV2, true);
-        //CameraReflectionApi.native_set("com.qti.node.eisv2","1");
-        //Log.d(TAG,"Points:"+captureBuilder.get(TONEMAP_PRESET_CURVE));
     }
-
-    void put(int in) {
-        sharedPreferencesEditor.putInt(mCameraID+"Settings:" + count, in);
+    void put(int in,String name) {
+        Log.d(TAG,"Saved "+name+":"+in);
+        sharedPreferencesEditor.putInt("ID"+String.format("%03d",Integer.parseInt(mCameraID))+"_i_"+name, in);
         count++;
     }
 
-    void put(double in) {
-        sharedPreferencesEditor.putFloat(mCameraID+"Settings:" + count, (float) in);
+    void put(double in,String name) {
+        Log.d(TAG,"Saved "+name+":"+in);
+        sharedPreferencesEditor.putFloat("ID"+String.format("%03d",Integer.parseInt(mCameraID))+"_d_"+name, (float) in);
         count++;
     }
 
-    void put(String in) {
-        sharedPreferencesEditor.putString(mCameraID+"Settings:" + count, in);
+    void put(String in,String name) {
+        Log.d(TAG,"Saved "+name+":"+in);
+        sharedPreferencesEditor.putString("ID"+String.format("%03d",Integer.parseInt(mCameraID))+"_s_"+name, in);
         count++;
     }
 
-    void put(boolean in) {
-        sharedPreferencesEditor.putBoolean(mCameraID+"Settings:" + count, in);
+    void put(boolean in,String name) {
+        Log.d(TAG,"Saved "+name+":"+in);
+        sharedPreferencesEditor.putBoolean("ID"+String.format("%03d",Integer.parseInt(mCameraID))+"_b_"+name, in);
         count++;
     }
 
-    boolean get(boolean in) {
-        boolean result = sharedPreferences.getBoolean(mCameraID+"Settings:" + count, in);
+    boolean get(boolean in,String name) {
+        boolean result = sharedPreferences.getBoolean("ID"+String.format("%03d",Integer.parseInt(mCameraID))+"_b_"+name, in);
+        Log.d(TAG,"Loaded "+name+":"+result);
         count++;
         return result;
     }
 
-    int get(int cur) {
+    int get(int cur,String name) {
         int result;
-        result = sharedPreferences.getInt(mCameraID+"Settings:" + count, cur);
+        result = sharedPreferences.getInt("ID"+String.format("%03d",Integer.parseInt(mCameraID))+"_i_"+name, cur);
+        Log.d(TAG,"Loaded "+name+":"+result);
         count++;
         return result;
     }
 
-    double get(double cur) {
+    double get(double cur,String name) {
         double result;
-        result = sharedPreferences.getFloat(mCameraID+"Settings:" + count, (float) (cur));
+        result = sharedPreferences.getFloat("ID"+String.format("%03d",Integer.parseInt(mCameraID))+"_d_"+name, (float) (cur));
+        Log.d(TAG,"Loaded "+name+":"+result);
         count++;
         return result;
     }
 
-    String get(String cur) {
+    String get(String cur,String name) {
         String result;
-        result = (sharedPreferences.getString(mCameraID+"Settings:" + count, (cur)));
+        result = (sharedPreferences.getString("ID"+String.format("%03d",Integer.parseInt(mCameraID))+"_s_"+name, (cur)));
+        Log.d(TAG,"Loaded "+name+":"+result);
         count++;
         return result;
+    }
+    public void ExportSettings(){
+       save();
+       Map<String, ?> allKeys =  sharedPreferences.getAll();
+       File configFile = new File(Environment.getExternalStorageDirectory()+"//DCIM//PhotonCamera//Settings.ini");
+       Properties props = new Properties();
+       try {
+           if (!configFile.exists()) configFile.createNewFile();
+           props.load(new FileInputStream(configFile));
+       }
+       catch(IOException e){
+           e.printStackTrace();
+           return;
+       }
+        for (Map.Entry<String, ?> entry : allKeys.entrySet()) {
+           props.setProperty(entry.getKey(),entry.getValue().toString());
+           Log.v(TAG,"setProperty:"+entry.getKey()+" = "+entry.getValue().toString());
+       }
+
+        try {
+            props.store(new FileOutputStream(configFile),"PhotonCamera settings file");
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+    public void ImportSettings(){
+        Map<String, ?> allKeys =  sharedPreferences.getAll();
+        File configFile = new File(Environment.getExternalStorageDirectory()+"//DCIM//PhotonCamera//Settings.ini");
+        Properties props = new Properties();
+        try {
+            if (!configFile.exists()) configFile.createNewFile();
+            props.load(new FileInputStream(configFile));
+        }
+        catch(IOException e){
+            e.printStackTrace();
+            return;
+        }
+        mCameraID = props.getProperty("Camera");
+        String ids = props.getProperty("Cameras",CameraManager2.cameraManager2.getCameraIdList().toString());
+        ids = ids.replace("]","");
+        ids = ids.replace("[","");
+        ids = ids.replace(" ","");
+        String idarr[] = ids.split(",");
+        CameraManager2.cameraManager2.mCameraIDs.clear();
+        for(String id : idarr) {
+            Log.d(TAG,"CameraID:"+id);
+            CameraManager2.cameraManager2.mCameraIDs.add(id);
+        }
+        for (Map.Entry<Object, Object> entry : props.entrySet()){
+            String property = entry.getValue().toString();
+            String name = entry.getKey().toString();
+            if(name.charAt(5) == '_' && name.length() > 6) {
+                Log.d(TAG,"Importing:"+name.charAt(6)+" "+name);
+                switch (name.charAt(6)) {
+                    case 'b':
+                        sharedPreferencesEditor.putBoolean(name, Boolean.parseBoolean(property));
+                        break;
+                    case 'i':
+                        sharedPreferencesEditor.putInt(name, Integer.parseInt(property));
+                        break;
+                    case 'd':
+                        sharedPreferencesEditor.putFloat(name, Float.parseFloat(property));
+                        break;
+                    case 's':
+                        sharedPreferencesEditor.putString(name, property);
+                        break;
+                }
+            }
+        }
+        load();
+        Interface.i.settingsActivity.set();
     }
 }
