@@ -16,6 +16,7 @@ import com.eszdman.photoncamera.R;
 import com.eszdman.photoncamera.api.Camera2ApiAutoFix;
 import com.eszdman.photoncamera.api.Interface;
 import com.eszdman.photoncamera.api.Permissions;
+import com.eszdman.photoncamera.util.FileManager;
 
 import org.opencv.android.OpenCVLoader;
 
@@ -37,30 +38,8 @@ public class MainActivity extends AppCompatActivity {
     }
     @Override
     public void onBackPressed() {
-        /*if (CameraFragment.context.mState != 5) {
-            super.onBackPressed();
-            return;
-        }
-        Intent intent = this.getIntent();
-        this.finish();
-        this.startActivity(intent);
-        */
         super.onBackPressed();
         Animatoo.animateShrink(this);
-    }
-
-    ImageView grid;
-    public void onCameraResume(){
-        Interface.i.swipedetection.RunDetection();
-    }
-    public static void onCameraViewCreated(){
-        Interface.i.manual = new Manual();
-
-        //Interface.i.swipedetection.RunDetection();
-    }
-    public static void onCameraInitialization(){
-        Camera2ApiAutoFix.Init();
-        Interface.i.manual.Init();
     }
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -68,9 +47,9 @@ public class MainActivity extends AppCompatActivity {
         getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,WindowManager.LayoutParams.FLAG_FULLSCREEN);
         act = this;
         Interface inter = new Interface(this);
-//        Wrapper.Test();
+        //Wrapper.Test();
         Permissions.RequestPermissions(this, 2, new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE, Manifest.permission.CAMERA});
-
+        FileManager.CreateFolders();
         CameraFragment.context = CameraFragment.newInstance();
         inter.camera = CameraFragment.context;
         setContentView(R.layout.activity_camera);
